move api to route
