+++ server/model/index.js
@@ -1,0 +1,1 @@
+// mongoose model
